REFACTOR: moving setup of server, berlin results and homepage to their own files
