+++ src/go_files/web_pages/berlin.go
@@ -1,0 +1,21 @@
+package fffwebpages
+
+import (
+  "log"
+  "net/http"
+  "io/ioutil"
+)
+
+
+
+// handles requests to the forecast page
+func PresentBerlinFlightDestinations(w http.ResponseWriter, r *http.Request) {
+	// serving a static file
+	pageContent, err := ioutil.ReadFile("src/html/berlin-flight-destinations.html")
+	if err != nil {
+		log.Printf("Error reading forecast page file: %v", err)
+		http.Error(w, "Internal server error", 500)
+		return
+	}
+	w.Write(pageContent)
+}
